move pipeline metrics out
