+++ .changeset/local-video-subscriber-timing.md
@@ -1,0 +1,5 @@
+---
+libwebrtc: patch
+---
+
+Keep one-frame native video streams as latest-frame queues. Move local video subscriber timing metrics into the example and use the WGPU paint callback as the render boundary.
