+++ examples/local_video/src/subscriber.rs
@@ -13,21 +13,22 @@ use livekit_api::access_token;
 use log::{debug, info};
 use parking_lot::Mutex;
 use std::{
-    collections::{HashMap, VecDeque},
+    collections::HashMap,
     env,
     sync::OnceLock,
     sync::{
         atomic::{AtomicBool, AtomicU32, AtomicU64, Ordering},
         Arc,
     },
-    thread,
     time::{Duration, Instant, SystemTime, UNIX_EPOCH},
 };
 
 mod codec_display;
+mod subscriber_timing;
 mod viewport_aspect;
 
 use codec_display::{codec_from_mime, codec_with_implementation};
+use subscriber_timing::SubscriberTimingHandle;
 use viewport_aspect::AspectConstrainedViewport;
 
 #[cfg(target_os = "macos")]
@@ -403,26 +404,6 @@ struct Args {
     #[arg(long)]
     display_timestamp: bool,
 
-    /// Add a per-frame empty command buffer to measure post-submit GPU completion timing
-    #[arg(long)]
-    gpu_completion_probe: bool,
-
-    /// Hide on-screen overlays and controls so the render path only paints video
-    #[arg(long)]
-    no_overlay: bool,
-
-    /// Disable periodic WebRTC stats polling while keeping frame latency logs
-    #[arg(long)]
-    no_stats: bool,
-
-    /// Initial window long edge in pixels
-    #[arg(long)]
-    window_long_edge: Option<f32>,
-
-    /// Only hand every Nth decoded frame to the renderer
-    #[arg(long)]
-    render_frame_step: Option<u32>,
-
     /// Shared encryption key for E2EE (enables AES-GCM end-to-end encryption when set; must match publisher's key)
     #[arg(long)]
     e2ee_key: Option<String>,
@@ -438,14 +419,6 @@ struct SharedYuv {
 
 struct LatestRenderFrameSlot {
     frame: Mutex<Option<BoxVideoFrame>>,
-}
-
-fn effective_render_frame_step(
-    render_frame_step: Option<u32>,
-    _no_overlay: bool,
-    _no_stats: bool,
-) -> u32 {
-    render_frame_step.unwrap_or(1).max(1)
 }
 
 impl LatestRenderFrameSlot {
@@ -489,38 +462,9 @@ impl AtomicVideoSize {
     }
 }
 
+/// Carried from prepare into the WGPU paint callback to stamp the paint boundary.
 #[derive(Clone, Copy, Debug)]
-struct SubscriberTimingSample {
-    frame_id: Option<u32>,
-    sensor_exposure_timestamp_us: u64,
-    webrtc_receive_timestamp_us: Option<u64>,
-    decoder_upload_timestamp_us: Option<u64>,
-    decoder_output_timestamp_us: Option<u64>,
-    frame_sink_timestamp_us: Option<u64>,
-    frame_prepare_timestamp_us: Option<u64>,
-    frame_painted_timestamp_us: Option<u64>,
-    frame_uploaded_to_gpu_timestamp_us: Option<u64>,
-}
-
-impl SubscriberTimingSample {
-    fn new(sensor_exposure_timestamp_us: u64, frame_id: Option<u32>) -> Self {
-        Self {
-            frame_id,
-            sensor_exposure_timestamp_us,
-            webrtc_receive_timestamp_us: None,
-            decoder_upload_timestamp_us: None,
-            decoder_output_timestamp_us: None,
-            frame_sink_timestamp_us: None,
-            frame_prepare_timestamp_us: None,
-            frame_painted_timestamp_us: None,
-            frame_uploaded_to_gpu_timestamp_us: None,
-        }
-    }
-}
-
-/// Carried from upload into the wgpu submit callback to stamp GPU upload completion.
-#[derive(Clone, Copy, Debug)]
-struct PendingGpuSample {
+struct PendingPaintSample {
     frame_id: Option<u32>,
     capture_timestamp_us: u64,
     prepare_timestamp_us: u64,
@@ -544,7 +488,7 @@ impl PendingPaintSampleSlot {
         }
     }
 
-    fn store(&self, sample: PendingGpuSample) {
+    fn store(&self, sample: PendingPaintSample) {
         let frame_id = sample.frame_id.unwrap_or(Self::NO_FRAME_ID);
         self.frame_id.store(frame_id, Ordering::Relaxed);
         self.prepare_timestamp_us.store(sample.prepare_timestamp_us, Ordering::Relaxed);
@@ -555,7 +499,7 @@ impl PendingPaintSampleSlot {
         self.capture_timestamp_us.store(Self::NO_SAMPLE, Ordering::Release);
     }
 
-    fn take(&self) -> Option<PendingGpuSample> {
+    fn take(&self) -> Option<PendingPaintSample> {
         let capture_timestamp_us =
             self.capture_timestamp_us.swap(Self::NO_SAMPLE, Ordering::Acquire);
         if capture_timestamp_us == Self::NO_SAMPLE {
@@ -566,616 +510,11 @@ impl PendingPaintSampleSlot {
             Self::NO_FRAME_ID => None,
             frame_id => Some(frame_id),
         };
-        Some(PendingGpuSample {
+        Some(PendingPaintSample {
             frame_id,
             capture_timestamp_us,
             prepare_timestamp_us: self.prepare_timestamp_us.load(Ordering::Relaxed),
         })
-    }
-}
-
-const MAX_SUBSCRIBER_TIMING_SAMPLES: usize = 300;
-const SUBSCRIBER_TIMING_DISPLAY_UPDATE_INTERVAL: Duration = Duration::from_millis(100);
-
-#[derive(Default)]
-struct SubscriberTimingState {
-    samples: HashMap<u64, SubscriberTimingSample>,
-    order: VecDeque<u64>,
-    latest_display_sample: Option<SubscriberTimingSample>,
-    render_latency_window: RenderLatencyWindow,
-    log_gpu_completion: bool,
-    displayed_timing_deltas: Option<SubscriberTimingDeltaValues>,
-    displayed_exp2recv_latency: Option<String>,
-    displayed_receive_to_render_latency: Option<String>,
-    displayed_e2e_latency: Option<String>,
-    last_latency_update: Option<Instant>,
-}
-
-impl SubscriberTimingState {
-    fn new(log_gpu_completion: bool) -> Self {
-        Self { log_gpu_completion, ..Self::default() }
-    }
-}
-
-#[derive(Default)]
-struct RenderLatencyWindow {
-    count: u64,
-    receive_to_decode_count: u64,
-    receive_to_decode_sum_us: u128,
-    receive_to_decode_min_us: Option<u64>,
-    receive_to_decode_max_us: Option<u64>,
-    decode_to_prepare_count: u64,
-    decode_to_prepare_sum_us: u128,
-    decode_to_prepare_min_us: Option<u64>,
-    decode_to_prepare_max_us: Option<u64>,
-    decode_to_sink_count: u64,
-    decode_to_sink_sum_us: u128,
-    decode_to_sink_min_us: Option<u64>,
-    decode_to_sink_max_us: Option<u64>,
-    sink_to_prepare_count: u64,
-    sink_to_prepare_sum_us: u128,
-    sink_to_prepare_min_us: Option<u64>,
-    sink_to_prepare_max_us: Option<u64>,
-    prepare_to_gpu_count: u64,
-    prepare_to_gpu_sum_us: u128,
-    prepare_to_gpu_min_us: Option<u64>,
-    prepare_to_gpu_max_us: Option<u64>,
-    prepare_to_paint_count: u64,
-    prepare_to_paint_sum_us: u128,
-    prepare_to_paint_min_us: Option<u64>,
-    prepare_to_paint_max_us: Option<u64>,
-    paint_to_gpu_count: u64,
-    paint_to_gpu_sum_us: u128,
-    paint_to_gpu_min_us: Option<u64>,
-    paint_to_gpu_max_us: Option<u64>,
-    decode_to_gpu_count: u64,
-    decode_to_gpu_sum_us: u128,
-    decode_to_gpu_min_us: Option<u64>,
-    decode_to_gpu_max_us: Option<u64>,
-    receive_to_paint_count: u64,
-    receive_to_paint_sum_us: u128,
-    receive_to_paint_min_us: Option<u64>,
-    receive_to_paint_max_us: Option<u64>,
-    receive_to_gpu_count: u64,
-    receive_to_gpu_sum_us: u128,
-    receive_to_gpu_min_us: Option<u64>,
-    receive_to_gpu_max_us: Option<u64>,
-    e2e_sum_us: u128,
-    e2e_min_us: Option<u64>,
-    e2e_max_us: Option<u64>,
-    last_log: Option<Instant>,
-}
-
-impl RenderLatencyWindow {
-    fn record(&mut self, sample: SubscriberTimingSample, now: Instant) {
-        let Some(frame_rendered_timestamp_us) =
-            sample.frame_uploaded_to_gpu_timestamp_us.or(sample.frame_painted_timestamp_us)
-        else {
-            return;
-        };
-
-        if let Some(webrtc_receive_timestamp_us) = sample.webrtc_receive_timestamp_us {
-            if let Some(decoder_output_timestamp_us) = sample.decoder_output_timestamp_us {
-                let latency_us =
-                    decoder_output_timestamp_us.saturating_sub(webrtc_receive_timestamp_us);
-                self.receive_to_decode_sum_us += u128::from(latency_us);
-                self.receive_to_decode_min_us = Some(
-                    self.receive_to_decode_min_us.map_or(latency_us, |min| min.min(latency_us)),
-                );
-                self.receive_to_decode_max_us = Some(
-                    self.receive_to_decode_max_us.map_or(latency_us, |max| max.max(latency_us)),
-                );
-                self.receive_to_decode_count += 1;
-            }
-
-            if let Some(frame_uploaded_to_gpu_timestamp_us) =
-                sample.frame_uploaded_to_gpu_timestamp_us
-            {
-                let latency_us =
-                    frame_uploaded_to_gpu_timestamp_us.saturating_sub(webrtc_receive_timestamp_us);
-                self.receive_to_gpu_sum_us += u128::from(latency_us);
-                self.receive_to_gpu_min_us =
-                    Some(self.receive_to_gpu_min_us.map_or(latency_us, |min| min.min(latency_us)));
-                self.receive_to_gpu_max_us =
-                    Some(self.receive_to_gpu_max_us.map_or(latency_us, |max| max.max(latency_us)));
-                self.receive_to_gpu_count += 1;
-            }
-
-            if let Some(frame_painted_timestamp_us) = sample.frame_painted_timestamp_us {
-                let latency_us =
-                    frame_painted_timestamp_us.saturating_sub(webrtc_receive_timestamp_us);
-                self.receive_to_paint_sum_us += u128::from(latency_us);
-                self.receive_to_paint_min_us = Some(
-                    self.receive_to_paint_min_us.map_or(latency_us, |min| min.min(latency_us)),
-                );
-                self.receive_to_paint_max_us = Some(
-                    self.receive_to_paint_max_us.map_or(latency_us, |max| max.max(latency_us)),
-                );
-                self.receive_to_paint_count += 1;
-            }
-        }
-
-        if let Some(decoder_output_timestamp_us) = sample.decoder_output_timestamp_us {
-            if let Some(frame_uploaded_to_gpu_timestamp_us) =
-                sample.frame_uploaded_to_gpu_timestamp_us
-            {
-                let latency_us =
-                    frame_uploaded_to_gpu_timestamp_us.saturating_sub(decoder_output_timestamp_us);
-                self.decode_to_gpu_sum_us += u128::from(latency_us);
-                self.decode_to_gpu_min_us =
-                    Some(self.decode_to_gpu_min_us.map_or(latency_us, |min| min.min(latency_us)));
-                self.decode_to_gpu_max_us =
-                    Some(self.decode_to_gpu_max_us.map_or(latency_us, |max| max.max(latency_us)));
-                self.decode_to_gpu_count += 1;
-            }
-        }
-
-        if let (Some(decoder_output_timestamp_us), Some(frame_prepare_timestamp_us)) =
-            (sample.decoder_output_timestamp_us, sample.frame_prepare_timestamp_us)
-        {
-            let latency_us = frame_prepare_timestamp_us.saturating_sub(decoder_output_timestamp_us);
-            self.decode_to_prepare_sum_us += u128::from(latency_us);
-            self.decode_to_prepare_min_us =
-                Some(self.decode_to_prepare_min_us.map_or(latency_us, |min| min.min(latency_us)));
-            self.decode_to_prepare_max_us =
-                Some(self.decode_to_prepare_max_us.map_or(latency_us, |max| max.max(latency_us)));
-            self.decode_to_prepare_count += 1;
-        }
-
-        if let (Some(decoder_output_timestamp_us), Some(frame_sink_timestamp_us)) =
-            (sample.decoder_output_timestamp_us, sample.frame_sink_timestamp_us)
-        {
-            let latency_us = frame_sink_timestamp_us.saturating_sub(decoder_output_timestamp_us);
-            self.decode_to_sink_sum_us += u128::from(latency_us);
-            self.decode_to_sink_min_us =
-                Some(self.decode_to_sink_min_us.map_or(latency_us, |min| min.min(latency_us)));
-            self.decode_to_sink_max_us =
-                Some(self.decode_to_sink_max_us.map_or(latency_us, |max| max.max(latency_us)));
-            self.decode_to_sink_count += 1;
-        }
-
-        if let (Some(frame_sink_timestamp_us), Some(frame_prepare_timestamp_us)) =
-            (sample.frame_sink_timestamp_us, sample.frame_prepare_timestamp_us)
-        {
-            let latency_us = frame_prepare_timestamp_us.saturating_sub(frame_sink_timestamp_us);
-            self.sink_to_prepare_sum_us += u128::from(latency_us);
-            self.sink_to_prepare_min_us =
-                Some(self.sink_to_prepare_min_us.map_or(latency_us, |min| min.min(latency_us)));
-            self.sink_to_prepare_max_us =
-                Some(self.sink_to_prepare_max_us.map_or(latency_us, |max| max.max(latency_us)));
-            self.sink_to_prepare_count += 1;
-        }
-
-        if let Some(frame_prepare_timestamp_us) = sample.frame_prepare_timestamp_us {
-            if let Some(frame_uploaded_to_gpu_timestamp_us) =
-                sample.frame_uploaded_to_gpu_timestamp_us
-            {
-                let latency_us =
-                    frame_uploaded_to_gpu_timestamp_us.saturating_sub(frame_prepare_timestamp_us);
-                self.prepare_to_gpu_sum_us += u128::from(latency_us);
-                self.prepare_to_gpu_min_us =
-                    Some(self.prepare_to_gpu_min_us.map_or(latency_us, |min| min.min(latency_us)));
-                self.prepare_to_gpu_max_us =
-                    Some(self.prepare_to_gpu_max_us.map_or(latency_us, |max| max.max(latency_us)));
-                self.prepare_to_gpu_count += 1;
-            }
-
-            if let Some(frame_painted_timestamp_us) = sample.frame_painted_timestamp_us {
-                let latency_us =
-                    frame_painted_timestamp_us.saturating_sub(frame_prepare_timestamp_us);
-                self.prepare_to_paint_sum_us += u128::from(latency_us);
-                self.prepare_to_paint_min_us = Some(
-                    self.prepare_to_paint_min_us.map_or(latency_us, |min| min.min(latency_us)),
-                );
-                self.prepare_to_paint_max_us = Some(
-                    self.prepare_to_paint_max_us.map_or(latency_us, |max| max.max(latency_us)),
-                );
-                self.prepare_to_paint_count += 1;
-
-                if let Some(frame_uploaded_to_gpu_timestamp_us) =
-                    sample.frame_uploaded_to_gpu_timestamp_us
-                {
-                    let latency_us = frame_uploaded_to_gpu_timestamp_us
-                        .saturating_sub(frame_painted_timestamp_us);
-                    self.paint_to_gpu_sum_us += u128::from(latency_us);
-                    self.paint_to_gpu_min_us = Some(
-                        self.paint_to_gpu_min_us.map_or(latency_us, |min| min.min(latency_us)),
-                    );
-                    self.paint_to_gpu_max_us = Some(
-                        self.paint_to_gpu_max_us.map_or(latency_us, |max| max.max(latency_us)),
-                    );
-                    self.paint_to_gpu_count += 1;
-                }
-            }
-        }
-
-        let e2e_us =
-            frame_rendered_timestamp_us.saturating_sub(sample.sensor_exposure_timestamp_us);
-        self.e2e_sum_us += u128::from(e2e_us);
-        self.e2e_min_us = Some(self.e2e_min_us.map_or(e2e_us, |min| min.min(e2e_us)));
-        self.e2e_max_us = Some(self.e2e_max_us.map_or(e2e_us, |max| max.max(e2e_us)));
-        self.count += 1;
-
-        if self
-            .last_log
-            .map_or(true, |last_log| now.duration_since(last_log) >= Duration::from_secs(2))
-        {
-            self.log_and_reset(now);
-        }
-    }
-
-    fn log_and_reset(&mut self, now: Instant) {
-        if self.count == 0 {
-            self.last_log = Some(now);
-            return;
-        }
-
-        let receive_to_gpu_avg_us = if self.receive_to_gpu_min_us.is_some() {
-            Some((self.receive_to_gpu_sum_us / u128::from(self.receive_to_gpu_count)) as u64)
-        } else {
-            None
-        };
-        let decode_to_gpu_avg_us = if self.decode_to_gpu_min_us.is_some() {
-            Some((self.decode_to_gpu_sum_us / u128::from(self.decode_to_gpu_count)) as u64)
-        } else {
-            None
-        };
-        let decode_to_prepare_avg_us = if self.decode_to_prepare_min_us.is_some() {
-            Some((self.decode_to_prepare_sum_us / u128::from(self.decode_to_prepare_count)) as u64)
-        } else {
-            None
-        };
-        let decode_to_sink_avg_us = if self.decode_to_sink_min_us.is_some() {
-            Some((self.decode_to_sink_sum_us / u128::from(self.decode_to_sink_count)) as u64)
-        } else {
-            None
-        };
-        let sink_to_prepare_avg_us = if self.sink_to_prepare_min_us.is_some() {
-            Some((self.sink_to_prepare_sum_us / u128::from(self.sink_to_prepare_count)) as u64)
-        } else {
-            None
-        };
-        let receive_to_decode_avg_us = if self.receive_to_decode_min_us.is_some() {
-            Some((self.receive_to_decode_sum_us / u128::from(self.receive_to_decode_count)) as u64)
-        } else {
-            None
-        };
-        let prepare_to_gpu_avg_us = if self.prepare_to_gpu_min_us.is_some() {
-            Some((self.prepare_to_gpu_sum_us / u128::from(self.prepare_to_gpu_count)) as u64)
-        } else {
-            None
-        };
-        let prepare_to_paint_avg_us = if self.prepare_to_paint_min_us.is_some() {
-            Some((self.prepare_to_paint_sum_us / u128::from(self.prepare_to_paint_count)) as u64)
-        } else {
-            None
-        };
-        let paint_to_gpu_avg_us = if self.paint_to_gpu_min_us.is_some() {
-            Some((self.paint_to_gpu_sum_us / u128::from(self.paint_to_gpu_count)) as u64)
-        } else {
-            None
-        };
-        let receive_to_paint_avg_us = if self.receive_to_paint_min_us.is_some() {
-            Some((self.receive_to_paint_sum_us / u128::from(self.receive_to_paint_count)) as u64)
-        } else {
-            None
-        };
-        let e2e_avg_us = (self.e2e_sum_us / u128::from(self.count)) as u64;
-
-        if self.receive_to_gpu_count == 0 {
-            info!(
-                "Subscriber render latency: frames={}, receive_to_decode avg={} min={} max={}, decoder_to_sink avg={} min={} max={}, sink_to_prepare avg={} min={} max={}, decoder_to_prepare avg={} min={} max={}, prepare_to_paint avg={} min={} max={}, receive_to_paint avg={} min={} max={}, e2e avg={} min={} max={}",
-                self.count,
-                latency_log_value(receive_to_decode_avg_us),
-                latency_log_value(self.receive_to_decode_min_us),
-                latency_log_value(self.receive_to_decode_max_us),
-                latency_log_value(decode_to_sink_avg_us),
-                latency_log_value(self.decode_to_sink_min_us),
-                latency_log_value(self.decode_to_sink_max_us),
-                latency_log_value(sink_to_prepare_avg_us),
-                latency_log_value(self.sink_to_prepare_min_us),
-                latency_log_value(self.sink_to_prepare_max_us),
-                latency_log_value(decode_to_prepare_avg_us),
-                latency_log_value(self.decode_to_prepare_min_us),
-                latency_log_value(self.decode_to_prepare_max_us),
-                latency_log_value(prepare_to_paint_avg_us),
-                latency_log_value(self.prepare_to_paint_min_us),
-                latency_log_value(self.prepare_to_paint_max_us),
-                latency_log_value(receive_to_paint_avg_us),
-                latency_log_value(self.receive_to_paint_min_us),
-                latency_log_value(self.receive_to_paint_max_us),
-                latency_log_value(Some(e2e_avg_us)),
-                latency_log_value(self.e2e_min_us),
-                latency_log_value(self.e2e_max_us),
-            );
-        } else {
-            info!(
-                "Subscriber render latency: frames={}, receive_to_decode avg={} min={} max={}, decoder_to_sink avg={} min={} max={}, sink_to_prepare avg={} min={} max={}, decoder_to_prepare avg={} min={} max={}, prepare_to_paint avg={} min={} max={}, paint_to_gpu avg={} min={} max={}, prepare_to_gpu avg={} min={} max={}, decoder_to_gpu avg={} min={} max={}, receive_to_paint avg={} min={} max={}, receive_to_gpu avg={} min={} max={}, e2e avg={} min={} max={}",
-                self.count,
-                latency_log_value(receive_to_decode_avg_us),
-                latency_log_value(self.receive_to_decode_min_us),
-                latency_log_value(self.receive_to_decode_max_us),
-                latency_log_value(decode_to_sink_avg_us),
-                latency_log_value(self.decode_to_sink_min_us),
-                latency_log_value(self.decode_to_sink_max_us),
-                latency_log_value(sink_to_prepare_avg_us),
-                latency_log_value(self.sink_to_prepare_min_us),
-                latency_log_value(self.sink_to_prepare_max_us),
-                latency_log_value(decode_to_prepare_avg_us),
-                latency_log_value(self.decode_to_prepare_min_us),
-                latency_log_value(self.decode_to_prepare_max_us),
-                latency_log_value(prepare_to_paint_avg_us),
-                latency_log_value(self.prepare_to_paint_min_us),
-                latency_log_value(self.prepare_to_paint_max_us),
-                latency_log_value(paint_to_gpu_avg_us),
-                latency_log_value(self.paint_to_gpu_min_us),
-                latency_log_value(self.paint_to_gpu_max_us),
-                latency_log_value(prepare_to_gpu_avg_us),
-                latency_log_value(self.prepare_to_gpu_min_us),
-                latency_log_value(self.prepare_to_gpu_max_us),
-                latency_log_value(decode_to_gpu_avg_us),
-                latency_log_value(self.decode_to_gpu_min_us),
-                latency_log_value(self.decode_to_gpu_max_us),
-                latency_log_value(receive_to_paint_avg_us),
-                latency_log_value(self.receive_to_paint_min_us),
-                latency_log_value(self.receive_to_paint_max_us),
-                latency_log_value(receive_to_gpu_avg_us),
-                latency_log_value(self.receive_to_gpu_min_us),
-                latency_log_value(self.receive_to_gpu_max_us),
-                latency_log_value(Some(e2e_avg_us)),
-                latency_log_value(self.e2e_min_us),
-                latency_log_value(self.e2e_max_us),
-            );
-        }
-
-        *self = Self { last_log: Some(now), ..Self::default() };
-    }
-}
-
-#[derive(Clone, Debug)]
-struct SubscriberTimingDeltaValues {
-    sensor_exposure: String,
-    webrtc_receive: String,
-    decoder_upload: String,
-    decoder_output: String,
-    frame_painted: String,
-    frame_uploaded_to_gpu: Option<String>,
-}
-
-impl SubscriberTimingDeltaValues {
-    fn from_sample(sample: SubscriberTimingSample) -> Self {
-        let base = sample.sensor_exposure_timestamp_us;
-        Self {
-            sensor_exposure: format_timing_delta_ms(base, base),
-            webrtc_receive: format_optional_timing_delta_ms(
-                sample.webrtc_receive_timestamp_us,
-                Some(base),
-            ),
-            decoder_upload: format_optional_timing_delta_ms(
-                sample.decoder_upload_timestamp_us,
-                sample.webrtc_receive_timestamp_us,
-            ),
-            decoder_output: format_optional_timing_delta_ms(
-                sample.decoder_output_timestamp_us,
-                sample.decoder_upload_timestamp_us,
-            ),
-            frame_painted: format_optional_timing_delta_ms(
-                sample.frame_painted_timestamp_us,
-                sample.decoder_output_timestamp_us,
-            ),
-            frame_uploaded_to_gpu: sample.frame_uploaded_to_gpu_timestamp_us.map(
-                |frame_uploaded_to_gpu_timestamp_us| {
-                    format_timing_delta_ms(
-                        frame_uploaded_to_gpu_timestamp_us,
-                        sample
-                            .frame_painted_timestamp_us
-                            .unwrap_or(frame_uploaded_to_gpu_timestamp_us),
-                    )
-                },
-            ),
-        }
-    }
-}
-
-struct SubscriberTimingOverlayValues {
-    deltas: SubscriberTimingDeltaValues,
-    exp2recv_latency: String,
-    receive_to_render_latency: String,
-    e2e_latency: String,
-}
-
-impl SubscriberTimingState {
-    fn record_subscribe_event(&mut self, event: SubscribeTimingEvent) {
-        if event.capture_timestamp_us == 0 {
-            return;
-        }
-
-        let updated_sample = {
-            let sample = self.get_or_insert_sample(event.capture_timestamp_us, event.frame_id);
-            match event.stage {
-                SubscribeTimingStage::WebrtcReceive => {
-                    sample.webrtc_receive_timestamp_us = Some(event.timestamp_us);
-                }
-                SubscribeTimingStage::DecoderUpload => {
-                    sample.decoder_upload_timestamp_us = Some(event.timestamp_us);
-                }
-                SubscribeTimingStage::DecoderOutput => {
-                    sample.decoder_output_timestamp_us = Some(event.timestamp_us);
-                }
-            }
-            *sample
-        };
-
-        if self
-            .latest_display_sample
-            .is_some_and(|sample| sample.sensor_exposure_timestamp_us == event.capture_timestamp_us)
-        {
-            self.latest_display_sample = Some(updated_sample);
-        }
-    }
-
-    fn record_frame_painted(
-        &mut self,
-        sensor_exposure_timestamp_us: u64,
-        frame_id: Option<u32>,
-        frame_prepare_timestamp_us: u64,
-        frame_painted_timestamp_us: u64,
-    ) {
-        let sample = self.get_or_insert_sample(sensor_exposure_timestamp_us, frame_id);
-        sample.frame_prepare_timestamp_us.get_or_insert(frame_prepare_timestamp_us);
-        sample.frame_painted_timestamp_us = Some(frame_painted_timestamp_us);
-        let sample = *sample;
-        self.latest_display_sample = Some(sample);
-        if !self.log_gpu_completion {
-            self.render_latency_window.record(sample, Instant::now());
-        }
-    }
-
-    fn record_frame_received_by_sink(
-        &mut self,
-        sensor_exposure_timestamp_us: u64,
-        frame_id: Option<u32>,
-        frame_sink_timestamp_us: u64,
-    ) {
-        let sample = self.get_or_insert_sample(sensor_exposure_timestamp_us, frame_id);
-        sample.frame_sink_timestamp_us = Some(frame_sink_timestamp_us);
-    }
-
-    fn record_frame_selected_for_render(
-        &mut self,
-        sensor_exposure_timestamp_us: u64,
-        frame_id: Option<u32>,
-        frame_selected_timestamp_us: u64,
-    ) {
-        let sample = self.get_or_insert_sample(sensor_exposure_timestamp_us, frame_id);
-        // The exact paint timestamp is only known inside the WGPU paint callback, after
-        // egui has already built this frame's overlay text. Use selection time as the
-        // display value for this pass; `record_frame_painted` overwrites it with the
-        // exact callback timestamp.
-        sample.frame_prepare_timestamp_us.get_or_insert(frame_selected_timestamp_us);
-        sample.frame_painted_timestamp_us.get_or_insert(frame_selected_timestamp_us);
-        let sample = *sample;
-        self.latest_display_sample = Some(sample);
-    }
-
-    fn record_frame_uploaded_to_gpu(
-        &mut self,
-        sensor_exposure_timestamp_us: u64,
-        frame_id: Option<u32>,
-        frame_prepare_timestamp_us: u64,
-        frame_uploaded_to_gpu_timestamp_us: u64,
-    ) -> SubscriberTimingSample {
-        let sample = self.get_or_insert_sample(sensor_exposure_timestamp_us, frame_id);
-        sample.frame_prepare_timestamp_us.get_or_insert(frame_prepare_timestamp_us);
-        sample.frame_uploaded_to_gpu_timestamp_us = Some(frame_uploaded_to_gpu_timestamp_us);
-        let sample = *sample;
-        self.latest_display_sample = Some(sample);
-        if self.log_gpu_completion {
-            self.render_latency_window.record(sample, Instant::now());
-        }
-        sample
-    }
-
-    fn display_sample(&self) -> Option<SubscriberTimingSample> {
-        self.latest_display_sample
-    }
-
-    fn display_overlay_lines(&mut self, now: Instant) -> Option<Vec<String>> {
-        let sample = self.display_sample()?;
-        let overlay_values = self.overlay_values(sample, now);
-        Some(build_timing_overlay_lines(sample, &overlay_values))
-    }
-
-    fn reset(&mut self) {
-        let log_gpu_completion = self.log_gpu_completion;
-        *self = Self::new(log_gpu_completion);
-    }
-
-    fn overlay_values(
-        &mut self,
-        sample: SubscriberTimingSample,
-        now: Instant,
-    ) -> SubscriberTimingOverlayValues {
-        let should_update = self.last_latency_update.map_or(true, |last_update| {
-            now.duration_since(last_update) >= SUBSCRIBER_TIMING_DISPLAY_UPDATE_INTERVAL
-        });
-
-        if should_update {
-            self.displayed_timing_deltas = Some(SubscriberTimingDeltaValues::from_sample(sample));
-            self.displayed_exp2recv_latency =
-                sample.webrtc_receive_timestamp_us.map(|webrtc_receive_timestamp_us| {
-                    format_latency_ms(
-                        webrtc_receive_timestamp_us,
-                        sample.sensor_exposure_timestamp_us,
-                    )
-                });
-            let frame_rendered_timestamp_us =
-                sample.frame_uploaded_to_gpu_timestamp_us.or(sample.frame_painted_timestamp_us);
-            self.displayed_receive_to_render_latency =
-                frame_rendered_timestamp_us.and_then(|frame_rendered_timestamp_us| {
-                    sample.webrtc_receive_timestamp_us.map(|webrtc_receive_timestamp_us| {
-                        format_latency_ms(frame_rendered_timestamp_us, webrtc_receive_timestamp_us)
-                    })
-                });
-            self.displayed_e2e_latency = frame_rendered_timestamp_us.map(|rendered_timestamp_us| {
-                format_latency_ms(rendered_timestamp_us, sample.sensor_exposure_timestamp_us)
-            });
-            self.last_latency_update = Some(now);
-        }
-
-        SubscriberTimingOverlayValues {
-            deltas: self
-                .displayed_timing_deltas
-                .clone()
-                .unwrap_or_else(|| SubscriberTimingDeltaValues::from_sample(sample)),
-            exp2recv_latency: self
-                .displayed_exp2recv_latency
-                .clone()
-                .unwrap_or_else(|| "NA".to_string()),
-            receive_to_render_latency: self
-                .displayed_receive_to_render_latency
-                .clone()
-                .unwrap_or_else(|| "NA".to_string()),
-            e2e_latency: self.displayed_e2e_latency.clone().unwrap_or_else(|| "NA".to_string()),
-        }
-    }
-
-    fn get_or_insert_sample(
-        &mut self,
-        sensor_exposure_timestamp_us: u64,
-        frame_id: Option<u32>,
-    ) -> &mut SubscriberTimingSample {
-        if !self.samples.contains_key(&sensor_exposure_timestamp_us) {
-            self.samples.insert(
-                sensor_exposure_timestamp_us,
-                SubscriberTimingSample::new(sensor_exposure_timestamp_us, frame_id),
-            );
-            self.order.push_back(sensor_exposure_timestamp_us);
-            self.prune();
-        }
-
-        let sample = self
-            .samples
-            .get_mut(&sensor_exposure_timestamp_us)
-            .expect("timing sample should exist after insertion");
-        if frame_id.is_some() {
-            sample.frame_id = frame_id;
-        }
-        sample
-    }
-
-    fn prune(&mut self) {
-        while self.order.len() > MAX_SUBSCRIBER_TIMING_SAMPLES {
-            if let Some(oldest) = self.order.pop_front() {
-                self.samples.remove(&oldest);
-                if self
-                    .latest_display_sample
-                    .is_some_and(|sample| sample.sensor_exposure_timestamp_us == oldest)
-                {
-                    self.latest_display_sample = None;
-                }
-            }
-        }
     }
 }
 
@@ -1391,49 +730,6 @@ fn current_timestamp_us() -> u64 {
     anchor.unix_timestamp_us.saturating_add(anchor.instant.elapsed().as_micros() as u64)
 }
 
-fn format_time_of_day_us(timestamp_us: u64) -> String {
-    let total_millis = timestamp_us / 1_000;
-    let millis = total_millis % 1_000;
-    let total_seconds = total_millis / 1_000;
-    let seconds = total_seconds % 60;
-    let minutes = (total_seconds / 60) % 60;
-    let hours = (total_seconds / 3_600) % 24;
-    format!("{hours:02}:{minutes:02}:{seconds:02}:{millis:03}")
-}
-
-fn format_timing_delta_ms(timestamp_us: u64, base_timestamp_us: u64) -> String {
-    let delta_us = i128::from(timestamp_us) - i128::from(base_timestamp_us);
-    if delta_us == 0 {
-        return "0.0ms".to_string();
-    }
-    format!("{:+.1}ms", delta_us as f64 / 1_000.0)
-}
-
-fn format_optional_timing_delta_ms(
-    timestamp_us: Option<u64>,
-    base_timestamp_us: Option<u64>,
-) -> String {
-    match (timestamp_us, base_timestamp_us) {
-        (Some(timestamp_us), Some(base_timestamp_us)) => {
-            format_timing_delta_ms(timestamp_us, base_timestamp_us)
-        }
-        _ => "+--.-ms".to_string(),
-    }
-}
-
-fn format_latency_ms(end_timestamp_us: u64, start_timestamp_us: u64) -> String {
-    end_timestamp_us
-        .checked_sub(start_timestamp_us)
-        .map_or_else(|| "NA".to_string(), |delta_us| format!("{:.1}ms", delta_us as f64 / 1_000.0))
-}
-
-fn latency_log_value(latency_us: Option<u64>) -> String {
-    latency_us.map_or_else(
-        || "NA".to_string(),
-        |latency_us| format!("{:.1}ms", latency_us as f64 / 1_000.0),
-    )
-}
-
 fn simulcast_state_full_dims(state: &Arc<Mutex<SimulcastState>>) -> Option<(u32, u32)> {
     let sc = state.lock();
     sc.full_dims
@@ -1455,160 +751,9 @@ fn video_status_line(
     }
 }
 
-const SUBSCRIBER_TIMING_LABEL_WIDTH: usize = 22;
-const SUBSCRIBER_TIMING_TIMESTAMP_WIDTH: usize = 12;
-const SUBSCRIBER_TIMING_DELTA_WIDTH: usize = 10;
-const SUBSCRIBER_TIMING_VALUE_WIDTH: usize =
-    SUBSCRIBER_TIMING_TIMESTAMP_WIDTH + 1 + SUBSCRIBER_TIMING_DELTA_WIDTH;
-const SUBSCRIBER_TIMING_LINE_WIDTH: usize =
-    SUBSCRIBER_TIMING_LABEL_WIDTH + 1 + SUBSCRIBER_TIMING_VALUE_WIDTH;
-
-fn subscriber_timing_label(label: &str) -> String {
-    format!("{label}:")
-}
-
-fn subscriber_timing_value_line(label: &str, value: &str) -> String {
-    let label = subscriber_timing_label(label);
-    format!(
-        "{label:<label_width$} {value:>value_width$}",
-        label_width = SUBSCRIBER_TIMING_LABEL_WIDTH,
-        value_width = SUBSCRIBER_TIMING_VALUE_WIDTH
-    )
-}
-
-fn subscriber_timing_line(label: &str, timestamp_us: Option<u64>, delta: &str) -> String {
-    let label = subscriber_timing_label(label);
-    match timestamp_us {
-        Some(timestamp_us) => format!(
-            "{label:<label_width$} {timestamp:>timestamp_width$} {delta:>delta_width$}",
-            timestamp = format_time_of_day_us(timestamp_us),
-            delta = delta,
-            label_width = SUBSCRIBER_TIMING_LABEL_WIDTH,
-            timestamp_width = SUBSCRIBER_TIMING_TIMESTAMP_WIDTH,
-            delta_width = SUBSCRIBER_TIMING_DELTA_WIDTH
-        ),
-        None => format!(
-            "{label:<label_width$} {timestamp:>timestamp_width$} {delta:>delta_width$}",
-            timestamp = "--:--:--:---",
-            delta = "+--.-ms",
-            label_width = SUBSCRIBER_TIMING_LABEL_WIDTH,
-            timestamp_width = SUBSCRIBER_TIMING_TIMESTAMP_WIDTH,
-            delta_width = SUBSCRIBER_TIMING_DELTA_WIDTH
-        ),
-    }
-}
-
-fn build_timing_overlay_lines(
-    sample: SubscriberTimingSample,
-    overlay_values: &SubscriberTimingOverlayValues,
-) -> Vec<String> {
-    let base = sample.sensor_exposure_timestamp_us;
-    let webrtc_receive = sample.webrtc_receive_timestamp_us;
-    let decoder_upload = sample.decoder_upload_timestamp_us;
-    let decoder_output = sample.decoder_output_timestamp_us;
-    let frame_painted = sample.frame_painted_timestamp_us;
-    let frame_uploaded_to_gpu = sample.frame_uploaded_to_gpu_timestamp_us;
-    let frame_id = sample.frame_id.map(|id| id.to_string()).unwrap_or_else(|| "NA".to_string());
-    let mut lines = vec![
-        subscriber_timing_value_line("Frame ID", &frame_id),
-        subscriber_timing_line(
-            "sensor exposure",
-            Some(base),
-            &overlay_values.deltas.sensor_exposure,
-        ),
-        subscriber_timing_line(
-            "webrtc receive",
-            webrtc_receive,
-            &overlay_values.deltas.webrtc_receive,
-        ),
-        subscriber_timing_line(
-            "decoder upload",
-            decoder_upload,
-            &overlay_values.deltas.decoder_upload,
-        ),
-        subscriber_timing_line(
-            "decoder output",
-            decoder_output,
-            &overlay_values.deltas.decoder_output,
-        ),
-        subscriber_timing_line(
-            "frame painted",
-            frame_painted,
-            &overlay_values.deltas.frame_painted,
-        ),
-    ];
-    if let Some(frame_uploaded_to_gpu_delta) = overlay_values.deltas.frame_uploaded_to_gpu.as_ref()
-    {
-        lines.push(subscriber_timing_line(
-            "GPU work done",
-            frame_uploaded_to_gpu,
-            frame_uploaded_to_gpu_delta,
-        ));
-    }
-    lines.extend([
-        subscriber_timing_value_line("Exposure to Receive", &overlay_values.exp2recv_latency),
-        subscriber_timing_value_line(
-            "Receive to Render",
-            &overlay_values.receive_to_render_latency,
-        ),
-        subscriber_timing_value_line("e2e latency", &overlay_values.e2e_latency),
-    ]);
-    lines
-}
-
-#[cfg(test)]
-fn assert_subscriber_timing_lines_are_stable(lines: &[String]) {
-    assert!(lines.iter().all(|line| line.len() == SUBSCRIBER_TIMING_LINE_WIDTH));
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    fn timestamp_us(hour: u64, minute: u64, second: u64, millisecond: u64) -> u64 {
-        (((hour * 3_600 + minute * 60 + second) * 1_000) + millisecond) * 1_000
-    }
-
-    fn subscribe_event(
-        stage: SubscribeTimingStage,
-        capture_timestamp_us: u64,
-        timestamp_us: u64,
-    ) -> SubscribeTimingEvent {
-        SubscribeTimingEvent { stage, timestamp_us, capture_timestamp_us, frame_id: Some(123) }
-    }
-
-    fn overlay_values(
-        sample: SubscriberTimingSample,
-        exp2recv_latency: &str,
-        receive_to_render_latency: &str,
-        e2e_latency: &str,
-    ) -> SubscriberTimingOverlayValues {
-        SubscriberTimingOverlayValues {
-            deltas: SubscriberTimingDeltaValues::from_sample(sample),
-            exp2recv_latency: exp2recv_latency.to_string(),
-            receive_to_render_latency: receive_to_render_latency.to_string(),
-            e2e_latency: e2e_latency.to_string(),
-        }
-    }
-
-    #[test]
-    fn low_latency_mode_defaults_to_rendering_every_decoded_frame() {
-        assert_eq!(effective_render_frame_step(None, true, true), 1);
-    }
-
-    #[test]
-    fn full_ui_modes_default_to_rendering_every_decoded_frame() {
-        assert_eq!(effective_render_frame_step(None, false, false), 1);
-        assert_eq!(effective_render_frame_step(None, true, false), 1);
-        assert_eq!(effective_render_frame_step(None, false, true), 1);
-    }
-
-    #[test]
-    fn explicit_render_frame_step_overrides_default_and_clamps_to_one() {
-        assert_eq!(effective_render_frame_step(Some(1), true, true), 1);
-        assert_eq!(effective_render_frame_step(Some(0), true, true), 1);
-        assert_eq!(effective_render_frame_step(Some(3), true, true), 3);
-    }
 
     #[test]
     fn subscriber_overlay_shows_status_without_timing() {
@@ -1621,201 +766,12 @@ mod tests {
         }));
         let simulcast =
             Arc::new(Mutex::new(SimulcastState { available: true, ..Default::default() }));
+        let subscriber_timing = SubscriberTimingHandle::new();
 
-        let lines = subscriber_overlay_lines(&shared, &simulcast, false, None)
+        let lines = subscriber_overlay_lines(&shared, &simulcast, false, &subscriber_timing)
             .expect("overlay should render");
 
         assert_eq!(lines, vec!["1280x720 29.6fps H264 NVDEC Simulcast"]);
-    }
-
-    #[test]
-    fn subscriber_timing_lines_match_requested_format() {
-        let base = timestamp_us(1, 2, 3, 456);
-        let sample = SubscriberTimingSample {
-            frame_id: Some(123),
-            sensor_exposure_timestamp_us: base,
-            webrtc_receive_timestamp_us: Some(base + 32_400),
-            decoder_upload_timestamp_us: Some(base + 35_500),
-            decoder_output_timestamp_us: Some(base + 55_300),
-            frame_sink_timestamp_us: Some(base + 55_900),
-            frame_prepare_timestamp_us: Some(base + 56_100),
-            frame_painted_timestamp_us: Some(base + 56_500),
-            frame_uploaded_to_gpu_timestamp_us: Some(base + 56_900),
-        };
-
-        let overlay_values = overlay_values(sample, "32.4ms", "24.5ms", "56.9ms");
-        let lines = build_timing_overlay_lines(sample, &overlay_values);
-        assert_subscriber_timing_lines_are_stable(&lines);
-        assert_eq!(
-            lines,
-            vec![
-                "Frame ID:                                  123",
-                "sensor exposure:       01:02:03:456      0.0ms",
-                "webrtc receive:        01:02:03:488    +32.4ms",
-                "decoder upload:        01:02:03:491     +3.1ms",
-                "decoder output:        01:02:03:511    +19.8ms",
-                "frame painted:         01:02:03:512     +1.2ms",
-                "GPU work done:         01:02:03:512     +0.4ms",
-                "Exposure to Receive:                    32.4ms",
-                "Receive to Render:                      24.5ms",
-                "e2e latency:                            56.9ms",
-            ]
-        );
-    }
-
-    #[test]
-    fn subscriber_timing_lines_use_placeholders_for_missing_stages() {
-        let base = timestamp_us(1, 2, 3, 456);
-        let sample = SubscriberTimingSample::new(base, None);
-
-        let overlay_values = overlay_values(sample, "NA", "NA", "NA");
-        let lines = build_timing_overlay_lines(sample, &overlay_values);
-        assert_subscriber_timing_lines_are_stable(&lines);
-        assert_eq!(
-            lines,
-            vec![
-                "Frame ID:                                   NA",
-                "sensor exposure:       01:02:03:456      0.0ms",
-                "webrtc receive:        --:--:--:---    +--.-ms",
-                "decoder upload:        --:--:--:---    +--.-ms",
-                "decoder output:        --:--:--:---    +--.-ms",
-                "frame painted:         --:--:--:---    +--.-ms",
-                "Exposure to Receive:                        NA",
-                "Receive to Render:                          NA",
-                "e2e latency:                                NA",
-            ]
-        );
-    }
-
-    #[test]
-    fn subscriber_latency_formatter_rejects_negative_latency() {
-        assert_eq!(format_latency_ms(900, 1_000), "NA");
-    }
-
-    #[test]
-    fn subscriber_timing_state_uses_selection_timestamp_until_paint_callback() {
-        let mut state = SubscriberTimingState::default();
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::WebrtcReceive,
-            1_000,
-            1_200,
-        ));
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::DecoderUpload,
-            1_000,
-            1_300,
-        ));
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::DecoderOutput,
-            1_000,
-            1_400,
-        ));
-
-        state.record_frame_selected_for_render(1_000, Some(123), 1_500);
-
-        let sample = state.display_sample().expect("selected frame should be displayable");
-        assert_eq!(sample.frame_id, Some(123));
-        assert_eq!(sample.webrtc_receive_timestamp_us, Some(1_200));
-        assert_eq!(sample.frame_prepare_timestamp_us, Some(1_500));
-        assert_eq!(sample.frame_painted_timestamp_us, Some(1_500));
-
-        let lines = state.display_overlay_lines(Instant::now()).expect("overlay should render");
-        assert_eq!(lines[5], "frame painted:         00:00:00:001     +0.1ms");
-    }
-
-    #[test]
-    fn subscriber_timing_state_displays_uploaded_sample() {
-        let mut state = SubscriberTimingState::default();
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::WebrtcReceive,
-            1_000,
-            1_200,
-        ));
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::DecoderUpload,
-            1_000,
-            1_300,
-        ));
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::DecoderOutput,
-            1_000,
-            1_400,
-        ));
-        assert!(state.display_sample().is_none());
-
-        state.record_frame_painted(1_000, Some(123), 1_450, 1_475);
-        let sample = state.record_frame_uploaded_to_gpu(1_000, Some(123), 1_450, 1_500);
-
-        assert_eq!(sample.frame_uploaded_to_gpu_timestamp_us, Some(1_500));
-        assert_eq!(sample.frame_prepare_timestamp_us, Some(1_450));
-        assert_eq!(sample.frame_painted_timestamp_us, Some(1_475));
-        assert_eq!(state.display_sample().unwrap().decoder_output_timestamp_us, Some(1_400));
-    }
-
-    #[test]
-    fn subscriber_timing_summary_latencies_refresh_at_ten_hz() {
-        let mut state = SubscriberTimingState::default();
-        let now = Instant::now();
-
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::WebrtcReceive,
-            1_000,
-            33_400,
-        ));
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::DecoderUpload,
-            1_000,
-            36_500,
-        ));
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::DecoderOutput,
-            1_000,
-            56_300,
-        ));
-        state.record_frame_painted(1_000, Some(1), 57_200, 57_900);
-        let lines = state.display_overlay_lines(now).expect("overlay should render");
-        assert_eq!(lines[3], "decoder upload:        00:00:00:036     +3.1ms");
-        assert_eq!(lines[4], "decoder output:        00:00:00:056    +19.8ms");
-        assert_eq!(lines[5], "frame painted:         00:00:00:057     +1.6ms");
-        assert_eq!(lines[6], "Exposure to Receive:                    32.4ms");
-        assert_eq!(lines[7], "Receive to Render:                      24.5ms");
-        assert_eq!(lines[8], "e2e latency:                            56.9ms");
-
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::WebrtcReceive,
-            1_000_000,
-            1_050_000,
-        ));
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::DecoderUpload,
-            1_000_000,
-            1_060_000,
-        ));
-        state.record_subscribe_event(subscribe_event(
-            SubscribeTimingStage::DecoderOutput,
-            1_000_000,
-            1_080_000,
-        ));
-        state.record_frame_painted(1_000_000, Some(2), 1_090_000, 1_100_000);
-        let lines = state
-            .display_overlay_lines(now + Duration::from_millis(99))
-            .expect("overlay should render");
-        assert_eq!(lines[3], "decoder upload:        00:00:01:060     +3.1ms");
-        assert_eq!(lines[4], "decoder output:        00:00:01:080    +19.8ms");
-        assert_eq!(lines[5], "frame painted:         00:00:01:100     +1.6ms");
-        assert_eq!(lines[6], "Exposure to Receive:                    32.4ms");
-        assert_eq!(lines[7], "Receive to Render:                      24.5ms");
-        assert_eq!(lines[8], "e2e latency:                            56.9ms");
-
-        let lines = state
-            .display_overlay_lines(now + Duration::from_millis(100))
-            .expect("overlay should render");
-        assert_eq!(lines[3], "decoder upload:        00:00:01:060    +10.0ms");
-        assert_eq!(lines[4], "decoder output:        00:00:01:080    +20.0ms");
-        assert_eq!(lines[5], "frame painted:         00:00:01:100    +20.0ms");
-        assert_eq!(lines[6], "Exposure to Receive:                    50.0ms");
-        assert_eq!(lines[7], "Receive to Render:                      50.0ms");
-        assert_eq!(lines[8], "e2e latency:                           100.0ms");
     }
 }
 
@@ -1831,10 +787,7 @@ async fn handle_track_subscribed(
     ctrl_c_received: &Arc<AtomicBool>,
     simulcast: &Arc<Mutex<SimulcastState>>,
     repaint_ctx: &Arc<OnceLock<egui::Context>>,
-    subscriber_timing_state: Option<Arc<Mutex<SubscriberTimingState>>>,
-    disable_stats: bool,
-    record_overlay_timing: bool,
-    render_frame_step: u32,
+    subscriber_timing: SubscriberTimingHandle,
 ) {
     // If a participant filter is set, skip others
     if let Some(ref allow) = allowed_identity {
@@ -1885,18 +838,16 @@ async fn handle_track_subscribed(
         s.codec = codec;
     }
 
-    let rtc_track = video_track.rtc_track();
-    if let Some(timing_state) = subscriber_timing_state.as_ref() {
-        let timing_state = timing_state.clone();
-        video_track.set_subscribe_timing_observer(Some(Arc::new(move |event| {
-            if !record_overlay_timing && event.stage == SubscribeTimingStage::DecoderUpload {
-                return;
-            }
-            timing_state.lock().record_subscribe_event(event);
-        })));
-    }
+    let mut timing_events = video_track.subscribe_timing_events();
+    let subscriber_timing_events = subscriber_timing.clone();
+    tokio::spawn(async move {
+        while let Some(event) = timing_events.next().await {
+            subscriber_timing_events.record_subscribe_event(event);
+        }
+    });
 
     // Start background sink task immediately so stats lookup cannot delay first-frame handling.
+    let rtc_track = video_track.rtc_track();
     let shared2 = shared.clone();
     let frame_slot_sink = frame_slot.clone();
     let video_size_sink = video_size.clone();
@@ -1904,8 +855,7 @@ async fn handle_track_subscribed(
     let my_sid = sid.clone();
     let ctrl_c_sink = ctrl_c_received.clone();
     let repaint_ctx_sink = repaint_ctx.clone();
-    let subscriber_timing_state_sink = subscriber_timing_state.clone();
-    let update_hud_state = !disable_stats || record_overlay_timing;
+    let subscriber_timing_sink = subscriber_timing.clone();
     // Initialize simulcast state for this publication
     {
         let mut sc = simulcast.lock();
@@ -1919,7 +869,6 @@ async fn handle_track_subscribed(
     tokio::spawn(async move {
         let mut sink = NativeVideoStream::new(rtc_track);
         let mut frames: u64 = 0;
-        let mut decoded_frames: u64 = 0;
         let mut last_log = Instant::now();
         let mut logged_first = false;
         let mut fps_window_frames: u64 = 0;
@@ -1938,19 +887,13 @@ async fn handle_track_subscribed(
             if drained_frames > 0 {
                 debug!("Dropped {drained_frames} stale decoded frames before render upload");
             }
-            decoded_frames += drained_frames + 1;
-            if render_frame_step > 1 && decoded_frames % u64::from(render_frame_step) != 0 {
-                continue;
-            }
             if let Some(metadata) = frame.frame_metadata {
                 if let Some(capture_timestamp_us) = metadata.user_timestamp {
-                    if let Some(timing_state) = subscriber_timing_state_sink.as_ref() {
-                        timing_state.lock().record_frame_received_by_sink(
-                            capture_timestamp_us,
-                            metadata.frame_id,
-                            current_timestamp_us(),
-                        );
-                    }
+                    subscriber_timing_sink.record_frame_received_by_sink(
+                        capture_timestamp_us,
+                        metadata.frame_id,
+                        current_timestamp_us(),
+                    );
                 }
             }
             let w = frame.buffer.width();
@@ -1962,33 +905,28 @@ async fn handle_track_subscribed(
             }
 
             let mut fps_update = None;
-            if !disable_stats {
-                // Update smoothed FPS (~500ms window)
-                fps_window_frames += 1;
-                let win_elapsed = fps_window_start.elapsed();
-                if win_elapsed >= Duration::from_millis(500) {
-                    let inst_fps =
-                        (fps_window_frames as f32) / (win_elapsed.as_secs_f32().max(0.001));
-                    fps_smoothed = if fps_smoothed <= 0.0 {
-                        inst_fps
-                    } else {
-                        // light EMA smoothing to reduce jitter
-                        (fps_smoothed * 0.7) + (inst_fps * 0.3)
-                    };
-                    fps_update = Some(fps_smoothed);
-                    fps_window_frames = 0;
-                    fps_window_start = Instant::now();
-                }
+            // Update smoothed FPS (~500ms window)
+            fps_window_frames += 1;
+            let win_elapsed = fps_window_start.elapsed();
+            if win_elapsed >= Duration::from_millis(500) {
+                let inst_fps = (fps_window_frames as f32) / (win_elapsed.as_secs_f32().max(0.001));
+                fps_smoothed = if fps_smoothed <= 0.0 {
+                    inst_fps
+                } else {
+                    // light EMA smoothing to reduce jitter
+                    (fps_smoothed * 0.7) + (inst_fps * 0.3)
+                };
+                fps_update = Some(fps_smoothed);
+                fps_window_frames = 0;
+                fps_window_start = Instant::now();
             }
 
-            if update_hud_state {
-                let mut s = shared2.lock();
-                if let Some(fps) = fps_update {
-                    s.fps = fps;
-                }
-                s.width = w;
-                s.height = h;
+            let mut s = shared2.lock();
+            if let Some(fps) = fps_update {
+                s.fps = fps;
             }
+            s.width = w;
+            s.height = h;
             video_size_sink.store(w, h);
             frame_slot_sink.store(frame);
 
@@ -1998,7 +936,7 @@ async fn handle_track_subscribed(
 
             frames += 1;
             let elapsed = last_log.elapsed();
-            if !disable_stats && elapsed >= Duration::from_secs(2) {
+            if elapsed >= Duration::from_secs(2) {
                 let fps = frames as f64 / elapsed.as_secs_f64();
                 info!("Receiving video: {}x{}, ~{:.1} fps", w, h, fps);
                 frames = 0;
@@ -2013,52 +951,50 @@ async fn handle_track_subscribed(
         }
     });
 
-    if !disable_stats {
-        let ctrl_c_stats = ctrl_c_received.clone();
-        let active_sid_stats = active_sid.clone();
-        let my_sid_stats = sid.clone();
-        let simulcast_stats = simulcast.clone();
-        let shared_stats = shared.clone();
-        tokio::spawn(async move {
-            let mut logged_initial = false;
-            let mut jitter_buffer_snapshot = None;
-            let mut last_jitter_buffer_log =
-                Instant::now().checked_sub(Duration::from_secs(5)).unwrap_or_else(Instant::now);
-            let mut interval = tokio::time::interval(Duration::from_secs(1));
-            interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
+    let ctrl_c_stats = ctrl_c_received.clone();
+    let active_sid_stats = active_sid.clone();
+    let my_sid_stats = sid.clone();
+    let simulcast_stats = simulcast.clone();
+    let shared_stats = shared.clone();
+    tokio::spawn(async move {
+        let mut logged_initial = false;
+        let mut jitter_buffer_snapshot = None;
+        let mut last_jitter_buffer_log =
+            Instant::now().checked_sub(Duration::from_secs(5)).unwrap_or_else(Instant::now);
+        let mut interval = tokio::time::interval(Duration::from_secs(1));
+        interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
 
-            loop {
-                if ctrl_c_stats.load(Ordering::Acquire) {
-                    break;
-                }
-                if active_sid_stats.lock().as_ref() != Some(&my_sid_stats) {
-                    break;
-                }
+        loop {
+            if ctrl_c_stats.load(Ordering::Acquire) {
+                break;
+            }
+            if active_sid_stats.lock().as_ref() != Some(&my_sid_stats) {
+                break;
+            }
 
-                match video_track.get_stats().await {
-                    Ok(stats) => {
-                        if !logged_initial {
-                            log_video_inbound_stats(&stats);
-                            logged_initial = true;
-                        }
-                        if last_jitter_buffer_log.elapsed() >= Duration::from_secs(5) {
-                            log_video_jitter_buffer_stats(&stats, &mut jitter_buffer_snapshot);
-                            last_jitter_buffer_log = Instant::now();
-                        }
-                        update_decoder_implementation_from_stats(&stats, &shared_stats);
-                        update_simulcast_quality_from_stats(&stats, &simulcast_stats);
-                    }
-                    Err(e) if !logged_initial => {
-                        debug!("Failed to get stats for video track: {:?}", e);
+            match video_track.get_stats().await {
+                Ok(stats) => {
+                    if !logged_initial {
+                        log_video_inbound_stats(&stats);
                         logged_initial = true;
                     }
-                    Err(_) => {}
+                    if last_jitter_buffer_log.elapsed() >= Duration::from_secs(5) {
+                        log_video_jitter_buffer_stats(&stats, &mut jitter_buffer_snapshot);
+                        last_jitter_buffer_log = Instant::now();
+                    }
+                    update_decoder_implementation_from_stats(&stats, &shared_stats);
+                    update_simulcast_quality_from_stats(&stats, &simulcast_stats);
                 }
-
-                interval.tick().await;
+                Err(e) if !logged_initial => {
+                    debug!("Failed to get stats for video track: {:?}", e);
+                    logged_initial = true;
+                }
+                Err(_) => {}
             }
-        });
-    }
+
+            interval.tick().await;
+        }
+    });
 }
 
 fn clear_hud_and_simulcast(
@@ -2066,7 +1002,7 @@ fn clear_hud_and_simulcast(
     frame_slot: &Arc<LatestRenderFrameSlot>,
     video_size: &Arc<AtomicVideoSize>,
     simulcast: &Arc<Mutex<SimulcastState>>,
-    subscriber_timing_state: Option<&Arc<Mutex<SubscriberTimingState>>>,
+    subscriber_timing: &SubscriberTimingHandle,
 ) {
     {
         let mut s = shared.lock();
@@ -2077,9 +1013,7 @@ fn clear_hud_and_simulcast(
         s.fps = 0.0;
     }
     frame_slot.clear();
-    if let Some(timing_state) = subscriber_timing_state {
-        timing_state.lock().reset();
-    }
+    subscriber_timing.reset();
     video_size.clear();
     let mut sc = simulcast.lock();
     *sc = SimulcastState::default();
@@ -2089,7 +1023,7 @@ fn subscriber_overlay_lines(
     shared: &Arc<Mutex<SharedYuv>>,
     simulcast: &Arc<Mutex<SimulcastState>>,
     include_timing: bool,
-    subscriber_timing_state: Option<&Arc<Mutex<SubscriberTimingState>>>,
+    subscriber_timing: &SubscriberTimingHandle,
 ) -> Option<Vec<String>> {
     let status_line = {
         let s = shared.lock();
@@ -2110,12 +1044,8 @@ fn subscriber_overlay_lines(
 
     let mut lines = vec![status_line];
     if include_timing {
-        if let Some(timing_state) = subscriber_timing_state {
-            if let Some(mut timing_lines) =
-                timing_state.lock().display_overlay_lines(Instant::now())
-            {
-                lines.append(&mut timing_lines);
-            }
+        if let Some(mut timing_lines) = subscriber_timing.display_overlay_lines(Instant::now()) {
+            lines.append(&mut timing_lines);
         }
     }
 
@@ -2133,7 +1063,7 @@ fn paint_subscriber_overlay(ctx: &egui::Context, lines: &[String]) {
                 .inner_margin(egui::Margin::same(6))
                 .show(ui, |ui| {
                     if lines.len() > 1 {
-                        ui.set_min_width(SUBSCRIBER_TIMING_LINE_WIDTH as f32 * 8.0);
+                        ui.set_min_width(subscriber_timing::TIMING_LINE_WIDTH as f32 * 8.0);
                     }
                     ui.add(
                         egui::Label::new(
@@ -2155,7 +1085,7 @@ fn handle_track_unsubscribed(
     video_size: &Arc<AtomicVideoSize>,
     active_sid: &Arc<Mutex<Option<TrackSid>>>,
     simulcast: &Arc<Mutex<SimulcastState>>,
-    subscriber_timing_state: Option<&Arc<Mutex<SubscriberTimingState>>>,
+    subscriber_timing: &SubscriberTimingHandle,
 ) {
     let sid = publication.sid().clone();
     let mut active = active_sid.lock();
@@ -2163,7 +1093,7 @@ fn handle_track_unsubscribed(
         info!("Video track unsubscribed ({}), clearing active sink", sid);
         *active = None;
     }
-    clear_hud_and_simulcast(shared, frame_slot, video_size, simulcast, subscriber_timing_state);
+    clear_hud_and_simulcast(shared, frame_slot, video_size, simulcast, subscriber_timing);
 }
 
 fn handle_track_unpublished(
@@ -2173,7 +1103,7 @@ fn handle_track_unpublished(
     video_size: &Arc<AtomicVideoSize>,
     active_sid: &Arc<Mutex<Option<TrackSid>>>,
     simulcast: &Arc<Mutex<SimulcastState>>,
-    subscriber_timing_state: Option<&Arc<Mutex<SubscriberTimingState>>>,
+    subscriber_timing: &SubscriberTimingHandle,
 ) {
     let sid = publication.sid().clone();
     let mut active = active_sid.lock();
@@ -2181,7 +1111,7 @@ fn handle_track_unpublished(
         info!("Video track unpublished ({}), clearing active sink", sid);
         *active = None;
     }
-    clear_hud_and_simulcast(shared, frame_slot, video_size, simulcast, subscriber_timing_state);
+    clear_hud_and_simulcast(shared, frame_slot, video_size, simulcast, subscriber_timing);
 }
 
 struct VideoApp {
@@ -2189,13 +1119,11 @@ struct VideoApp {
     frame_slot: Arc<LatestRenderFrameSlot>,
     video_size: Arc<AtomicVideoSize>,
     simulcast: Arc<Mutex<SimulcastState>>,
-    subscriber_timing_state: Option<Arc<Mutex<SubscriberTimingState>>>,
-    gpu_completion_probe: bool,
+    subscriber_timing: SubscriberTimingHandle,
     repaint_ctx: Arc<OnceLock<egui::Context>>,
     ctrl_c_received: Arc<AtomicBool>,
     viewport: AspectConstrainedViewport,
     display_timestamp: bool,
-    show_overlay: bool,
 }
 
 impl eframe::App for VideoApp {
@@ -2214,26 +1142,21 @@ impl eframe::App for VideoApp {
         if let Some(frame) = render_frame.as_ref() {
             if let Some(metadata) = frame.frame_metadata {
                 if let Some(capture_timestamp_us) = metadata.user_timestamp {
-                    if let Some(timing_state) = self.subscriber_timing_state.as_ref() {
-                        let frame_selected_timestamp_us = current_timestamp_us();
-                        timing_state.lock().record_frame_selected_for_render(
-                            capture_timestamp_us,
-                            metadata.frame_id,
-                            frame_selected_timestamp_us,
-                        );
-                    }
+                    self.subscriber_timing.record_frame_selected_for_render(
+                        capture_timestamp_us,
+                        metadata.frame_id,
+                        current_timestamp_us(),
+                    );
                 }
             }
         }
 
-        let overlay_lines = self.show_overlay.then(|| {
-            subscriber_overlay_lines(
-                &self.shared,
-                &self.simulcast,
-                self.display_timestamp,
-                self.subscriber_timing_state.as_ref(),
-            )
-        });
+        let overlay_lines = subscriber_overlay_lines(
+            &self.shared,
+            &self.simulcast,
+            self.display_timestamp,
+            &self.subscriber_timing,
+        );
 
         egui::CentralPanel::default().frame(egui::Frame::NONE).show(ctx, |ui| {
             ui.ctx().request_repaint();
@@ -2252,9 +1175,7 @@ impl eframe::App for VideoApp {
                         YuvPaintCallback {
                             render_frame: Mutex::new(render_frame),
                             video_size: self.video_size.clone(),
-                            subscriber_timing_state: self.subscriber_timing_state.clone(),
-                            gpu_completion_probe: self.gpu_completion_probe,
-                            ctrl_c_received: self.ctrl_c_received.clone(),
+                            subscriber_timing: self.subscriber_timing.clone(),
                         },
                     );
                     ui.painter().add(cb);
@@ -2262,40 +1183,38 @@ impl eframe::App for VideoApp {
             );
         });
 
-        if let Some(Some(lines)) = overlay_lines.as_ref() {
+        if let Some(lines) = overlay_lines.as_ref() {
             paint_subscriber_overlay(ctx, lines);
         }
 
-        if self.show_overlay {
-            // Simulcast layer controls: bottom-left overlay
-            egui::Area::new("simulcast_controls".into())
-                .anchor(egui::Align2::LEFT_BOTTOM, egui::vec2(10.0, -10.0))
-                .interactable(true)
-                .show(ctx, |ui| {
-                    let mut sc = self.simulcast.lock();
-                    if !sc.available {
-                        return;
-                    }
-                    let selected = sc.requested_quality.or(sc.active_quality);
-                    ui.horizontal(|ui| {
-                        let choices = [
-                            (livekit::track::VideoQuality::Low, "Low"),
-                            (livekit::track::VideoQuality::Medium, "Med"),
-                            (livekit::track::VideoQuality::High, "High"),
-                        ];
-                        for (q, label) in choices {
-                            let is_selected = selected.is_some_and(|s| s == q);
-                            let resp = ui.selectable_label(is_selected, label);
-                            if resp.clicked() {
-                                if let Some(ref pub_remote) = sc.publication {
-                                    pub_remote.set_video_quality(q);
-                                    sc.requested_quality = Some(q);
-                                }
+        // Simulcast layer controls: bottom-left overlay
+        egui::Area::new("simulcast_controls".into())
+            .anchor(egui::Align2::LEFT_BOTTOM, egui::vec2(10.0, -10.0))
+            .interactable(true)
+            .show(ctx, |ui| {
+                let mut sc = self.simulcast.lock();
+                if !sc.available {
+                    return;
+                }
+                let selected = sc.requested_quality.or(sc.active_quality);
+                ui.horizontal(|ui| {
+                    let choices = [
+                        (livekit::track::VideoQuality::Low, "Low"),
+                        (livekit::track::VideoQuality::Medium, "Med"),
+                        (livekit::track::VideoQuality::High, "High"),
+                    ];
+                    for (q, label) in choices {
+                        let is_selected = selected.is_some_and(|s| s == q);
+                        let resp = ui.selectable_label(is_selected, label);
+                        if resp.clicked() {
+                            if let Some(ref pub_remote) = sc.publication {
+                                pub_remote.set_video_quality(q);
+                                sc.requested_quality = Some(q);
                             }
                         }
-                    });
+                    }
                 });
-        }
+            });
 
         ctx.request_repaint_after(viewport_aspect::VIDEO_REPAINT_INTERVAL);
     }
@@ -2381,8 +1300,7 @@ async fn run(args: Args, ctrl_c_received: Arc<AtomicBool>) -> Result<()> {
     }));
     let frame_slot = Arc::new(LatestRenderFrameSlot::new());
     let video_size = Arc::new(AtomicVideoSize::default());
-    let subscriber_timing_state =
-        Arc::new(Mutex::new(SubscriberTimingState::new(args.gpu_completion_probe)));
+    let subscriber_timing = SubscriberTimingHandle::new();
 
     // Subscribe to room events: on first video track, start sink task
     let allowed_identity = args.participant.clone();
@@ -2397,18 +1315,10 @@ async fn run(args: Args, ctrl_c_received: Arc<AtomicBool>) -> Result<()> {
     let simulcast_events = simulcast.clone();
     let repaint_ctx_events = repaint_ctx.clone();
     let ctrl_c_events = ctrl_c_received.clone();
-    let subscriber_timing_state_events = Some(subscriber_timing_state.clone());
-    let disable_stats = args.no_stats;
-    let record_overlay_timing = !args.no_overlay;
-    let render_frame_step =
-        effective_render_frame_step(args.render_frame_step, args.no_overlay, args.no_stats);
-    if render_frame_step > 1 {
-        info!("Rendering every {render_frame_step} decoded frames");
-    }
+    let subscriber_timing_events = subscriber_timing.clone();
     tokio::spawn(async move {
         let active_sid = active_sid.clone();
         let simulcast = simulcast_events;
-        let subscriber_timing_state = subscriber_timing_state_events;
         let mut events = room.subscribe();
         info!("Subscribed to room events");
         while let Some(evt) = events.recv().await {
@@ -2427,10 +1337,7 @@ async fn run(args: Args, ctrl_c_received: Arc<AtomicBool>) -> Result<()> {
                         &ctrl_c_events,
                         &simulcast,
                         &repaint_ctx_events,
-                        subscriber_timing_state.clone(),
-                        disable_stats,
-                        record_overlay_timing,
-                        render_frame_step,
+                        subscriber_timing_events.clone(),
                     )
                     .await;
                 }
@@ -2442,7 +1349,7 @@ async fn run(args: Args, ctrl_c_received: Arc<AtomicBool>) -> Result<()> {
                         &video_size_events,
                         &active_sid,
                         &simulcast,
-                        subscriber_timing_state.as_ref(),
+                        &subscriber_timing_events,
                     );
                 }
                 RoomEvent::TrackUnpublished { publication, .. } => {
@@ -2453,7 +1360,7 @@ async fn run(args: Args, ctrl_c_received: Arc<AtomicBool>) -> Result<()> {
                         &video_size_events,
                         &active_sid,
                         &simulcast,
-                        subscriber_timing_state.as_ref(),
+                        &subscriber_timing_events,
                     );
                 }
                 _ => {}
@@ -2461,38 +1368,20 @@ async fn run(args: Args, ctrl_c_received: Arc<AtomicBool>) -> Result<()> {
         }
     });
 
-    let default_window_long_edge = if args.no_overlay && args.no_stats {
-        viewport_aspect::MIN_LONG_EDGE
-    } else {
-        viewport_aspect::DEFAULT_INITIAL_LONG_EDGE
-    };
-    let window_long_edge = args.window_long_edge.unwrap_or(default_window_long_edge);
-    let uses_default_window_long_edge =
-        (window_long_edge - viewport_aspect::DEFAULT_INITIAL_LONG_EDGE).abs() < f32::EPSILON;
-    let viewport = if uses_default_window_long_edge {
-        AspectConstrainedViewport::new(None)
-    } else {
-        AspectConstrainedViewport::with_initial_long_edge(None, window_long_edge)
-    };
+    let viewport = AspectConstrainedViewport::new(None);
     // Start UI
     let app = VideoApp {
         shared,
         frame_slot,
         video_size,
         simulcast,
-        subscriber_timing_state: Some(subscriber_timing_state),
-        gpu_completion_probe: args.gpu_completion_probe,
+        subscriber_timing,
         repaint_ctx,
         ctrl_c_received: ctrl_c_received.clone(),
         viewport,
         display_timestamp: args.display_timestamp,
-        show_overlay: !args.no_overlay,
     };
-    let native_options = if uses_default_window_long_edge {
-        viewport_aspect::native_options(None)
-    } else {
-        viewport_aspect::native_options_with_initial_long_edge(None, window_long_edge)
-    };
+    let native_options = viewport_aspect::native_options(None);
     eframe::run_native(
         "LiveKit Video Subscriber",
         native_options,
@@ -2510,9 +1399,7 @@ async fn run(args: Args, ctrl_c_received: Arc<AtomicBool>) -> Result<()> {
 struct YuvPaintCallback {
     render_frame: Mutex<Option<BoxVideoFrame>>,
     video_size: Arc<AtomicVideoSize>,
-    subscriber_timing_state: Option<Arc<Mutex<SubscriberTimingState>>>,
-    gpu_completion_probe: bool,
-    ctrl_c_received: Arc<AtomicBool>,
+    subscriber_timing: SubscriberTimingHandle,
 }
 
 struct YuvGpuState {
@@ -2533,7 +1420,6 @@ struct YuvGpuState {
     dims: (u32, u32),
     yuv_layout: u32,
     pending_paint_sample: PendingPaintSampleSlot,
-    gpu_completion_poller_started: bool,
     cpu_upload_logged: bool,
     #[cfg(target_os = "macos")]
     native_resources: Option<macos_native_video::NativeFrameResources>,
@@ -2614,30 +1500,6 @@ impl YuvGpuState {
         }
         queue.write_buffer(&self.params_buf, 0, bytemuck::bytes_of(&params));
         self.params = params;
-    }
-
-    fn ensure_gpu_completion_poller(
-        &mut self,
-        device: &wgpu::Device,
-        ctrl_c_received: Arc<AtomicBool>,
-    ) {
-        if self.gpu_completion_poller_started {
-            return;
-        }
-        self.gpu_completion_poller_started = true;
-
-        let device = device.clone();
-        let spawn_result =
-            thread::Builder::new().name("local-video-gpu-poll".to_string()).spawn(move || {
-                while !ctrl_c_received.load(Ordering::Acquire) {
-                    let _ = device.poll(wgpu::PollType::Poll);
-                    thread::sleep(Duration::from_micros(500));
-                }
-            });
-
-        if let Err(err) = spawn_result {
-            debug!("Unable to start GPU completion poller: {err}");
-        }
     }
 }
 
@@ -2849,7 +1711,6 @@ impl CallbackTrait for YuvPaintCallback {
                 dims: (0, 0),
                 yuv_layout: 0,
                 pending_paint_sample: PendingPaintSampleSlot::new(),
-                gpu_completion_poller_started: false,
                 cpu_upload_logged: false,
                 #[cfg(target_os = "macos")]
                 native_resources: None,
@@ -2863,15 +1724,12 @@ impl CallbackTrait for YuvPaintCallback {
             resources.insert(new_state);
         }
         let state = resources.get_mut::<YuvGpuState>().unwrap();
-        if self.gpu_completion_probe {
-            state.ensure_gpu_completion_poller(device, self.ctrl_c_received.clone());
-        }
 
         let frame_for_upload = self.render_frame.lock().take().map(|frame| {
             let prepare_timestamp_us = current_timestamp_us();
             let frame_id = frame.frame_metadata.and_then(|m| m.frame_id);
             let sample = frame.frame_metadata.and_then(|metadata| {
-                metadata.user_timestamp.map(|capture_timestamp_us| PendingGpuSample {
+                metadata.user_timestamp.map(|capture_timestamp_us| PendingPaintSample {
                     frame_id,
                     capture_timestamp_us,
                     prepare_timestamp_us,
@@ -2896,7 +1754,6 @@ impl CallbackTrait for YuvPaintCallback {
             state.recreate_bind_group(device);
         }
 
-        let mut gpu_sample_in_flight: Option<PendingGpuSample> = None;
         let mut frame_for_cpu_upload = frame_for_upload;
 
         #[cfg(target_os = "macos")]
@@ -2952,12 +1809,7 @@ impl CallbackTrait for YuvPaintCallback {
                                 },
                             );
                             match sample {
-                                Some(sample) => {
-                                    state.pending_paint_sample.store(sample);
-                                    if self.gpu_completion_probe {
-                                        gpu_sample_in_flight = Some(sample);
-                                    }
-                                }
+                                Some(sample) => state.pending_paint_sample.store(sample),
                                 None => state.pending_paint_sample.clear(),
                             }
                         }
@@ -3084,39 +1936,12 @@ impl CallbackTrait for YuvPaintCallback {
                 },
             );
             match sample {
-                Some(sample) => {
-                    state.pending_paint_sample.store(sample);
-                    if self.gpu_completion_probe {
-                        gpu_sample_in_flight = Some(sample);
-                    }
-                }
+                Some(sample) => state.pending_paint_sample.store(sample),
                 None => state.pending_paint_sample.clear(),
             }
         }
 
-        // Diagnostic only: ride an empty command buffer with egui's submit so we can
-        // stamp post-submit GPU completion without making it part of the default path.
-        if let Some(sample) = gpu_sample_in_flight {
-            let subscriber_timing_state = self.subscriber_timing_state.clone();
-            let encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                label: Some("yuv_gpu_done_probe"),
-            });
-            let cb = encoder.finish();
-            cb.on_submitted_work_done(move || {
-                if let Some(timing_state) = subscriber_timing_state.as_ref() {
-                    let frame_uploaded_to_gpu_timestamp_us = current_timestamp_us();
-                    timing_state.lock().record_frame_uploaded_to_gpu(
-                        sample.capture_timestamp_us,
-                        sample.frame_id,
-                        sample.prepare_timestamp_us,
-                        frame_uploaded_to_gpu_timestamp_us,
-                    );
-                }
-            });
-            vec![cb]
-        } else {
-            Vec::new()
-        }
+        Vec::new()
     }
 
     fn paint(
@@ -3139,15 +1964,12 @@ impl CallbackTrait for YuvPaintCallback {
         render_pass.draw(0..3, 0..1);
 
         if let Some(sample) = painted_sample {
-            if let Some(timing_state) = self.subscriber_timing_state.as_ref() {
-                let frame_painted_timestamp_us = current_timestamp_us();
-                timing_state.lock().record_frame_painted(
-                    sample.capture_timestamp_us,
-                    sample.frame_id,
-                    sample.prepare_timestamp_us,
-                    frame_painted_timestamp_us,
-                );
-            }
+            self.subscriber_timing.record_frame_painted(
+                sample.capture_timestamp_us,
+                sample.frame_id,
+                sample.prepare_timestamp_us,
+                current_timestamp_us(),
+            );
         }
     }
 }
