+++ libwebrtc/Cargo.toml
@@ -34,7 +34,6 @@ lazy_static = { workspace = true }
 parking_lot = { workspace = true }
 tokio = { workspace = true, default-features = false, features = ["sync", "macros"] }
 cxx = "1.0"
-futures-util = { workspace = true }
 rtrb = "0.3.3"
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
